refactoring, clean + mvp impl finish, di + rx added, app navigation added

+++ presentation/src/main/java/com/myst3ry/catchmovie/listener/OnMovieClickListener.java
@@ -2,5 +2,5 @@ package com.myst3ry.catchmovie.listener;
 
 public interface OnMovieClickListener {
 
-    void onMovieClick(final long movieId);
+    void onMovieClick(final int movieId);
 }
