+++ presentation/src/main/java/com/myst3ry/catchmovie/listener/OnTvShowClickListener.java
@@ -2,5 +2,5 @@ package com.myst3ry.catchmovie.listener;
 
 public interface OnTvShowClickListener {
 
-    void onTvShowClick(final long tvShowId);
+    void onTvShowClick(final int tvShowId);
 }
